add a common setup function for handler tests
